[유민형] Player 체력 UI 버그 수정

+++ 666GunsRemaster/Assets/YMH/Scripts/UIManager.cs
@@ -59,7 +59,7 @@ public class UIManager : MonoBehaviour
         }
     }
     //체력 게이지 UI
-    public void UpdatePlayerHealth(int maxHealth, int health)
+    public void UpdatePlayerHealth(float maxHealth, float health)
     {
         PlayerHealthSlider.value = health / maxHealth;
     }
